Fixed bump version pipeline
Now version is bumped on closed pull requests only and for limited set of branches.
If the version could not be bumped (e.g. tag already exists) the corresponding github action raises error,
which is now handled: if it occurs, version will not be bumped and pipeline will not fail.